added: REVIEW comment
Seems that a starting issue could be that the units for force in equation F=G*(2/d*d) could be wrong. Will keep taking a look later to see if something else could be wrong.

Also, if we are dealing with floats, G should perhaps also be a float instead of a double (but doesn't matter that much).